+++ Cargo.toml
@@ -48,6 +48,7 @@ members = [
   "crates/bitnet-eval-core",     # SRP: deterministic eval/scoring math helpers
   "crates/bitnet-feature-matrix",
   "crates/bitnet-server-health-types-core", # SRP: reusable server health endpoint contract types
+  "crates/bitnet-client-ip-core", # SRP: reusable client IP extraction from proxy headers
   "crates/bitnet-server",
   "crates/bitnet-cli-sampling-core", # SRP: reusable CLI sampling primitives
   "crates/bitnet-scoring-core", # SRP: reusable NLL/logit scoring primitives
@@ -149,6 +150,7 @@ default-members = [
   "crates/bitnet-cli",
   "crates/bitnet-scoring-core",
   "crates/bitnet-server",
+  "crates/bitnet-client-ip-core",
   "crates/bitnet-st2gguf",
   # Testing infrastructure crates (always built/tested by default)
   "crates/bitnet-bdd-grid",
