+++ crates/bitnet-client-ip-core/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "bitnet-client-ip-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "SRP client IP header parsing primitives for server middleware"
+rust-version.workspace = true
+
+[dependencies]
+http = "1.3.1"
+
+[lints]
+workspace = true
